UBUNTU SAUCE: ubuntu_blktrace_smoke_test: make sure $rd and $wr are not empty strings
If the test was not successfully executed, the grep command to calculate
the $rd and $wr value will fail.

In this case the following awk command will not be executed at all:

This will eventually casue errors like:
  ubuntu_blktrace_smoke_test.sh: line 237: [: -eq: unary operator expected
  ubuntu_blktrace_smoke_test.sh: line 240: [: -eq: unary operator expected

As the $\rd and $wr are empty string.

Fix this issue with arithmetic expansion to plus 0 to these two strings.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>