UBUNTU SAUCE: Add new misc event performance benchmark
The purpose of this test is to run this on a specific machine and
to measure various events.  The first cut of this test measures kernel
and userspace wakeups on an idle system using eventstat.

Each test scenario is run 5 times with an average/min/max/stddev CPU
values computed from the results with a threshold sanity check.  Data
can be scraped from the results log and pushed into a suitable database
trend analysis.

Typical output is as follows:

15:03:57 INFO | kernel_wakeups 4.450 5.720 5.670 4.980 4.960
15:03:57 INFO | kernel_wakeups_minimum 4.450
15:03:57 INFO | kernel_wakeups_maximum 5.720
15:03:57 INFO | kernel_wakeups_average 5.156
15:03:57 INFO | kernel_wakeups_maximum_error 24.631%
15:03:57 INFO | kernel_wakeups_stddev 0.536
15:03:57 INFO | kernel_wakeups_percent_stddev 10.400
15:03:57 INFO | FAIL: maximum error is greater than 5%
15:03:57 INFO |
15:09:02 INFO | userspace_wakeups 1385.130 1176.150 1700.970 1779.140 1350.720
15:09:02 INFO | userspace_wakeups_minimum 1176.150
15:09:02 INFO | userspace_wakeups_maximum 1779.140
15:09:02 INFO | userspace_wakeups_average 1478.422
15:09:02 INFO | userspace_wakeups_maximum_error 40.786%
15:09:02 INFO | userspace_wakeups_stddev 253.149
15:09:02 INFO | userspace_wakeups_percent_stddev 17.123
15:09:02 INFO | FAIL: maximum error is greater than 5%

Signed-off-by: Colin Ian King <colin.king@canonical.com>