UBUNTU SAUCE: Add new cyclic latenct tests performance metrics test
The purpose of this test is to run this on a specific machine and
to measure the latencies of various stress-ng stressors with the
stress-ng cyclic measuring test.  4 latency value per stress test
are produced: mean, mode, min, max

Typical output is as follows:

16:17:40 INFO | latency_cpu_mean_average 4704.750
16:17:40 INFO | latency_cpu_mode_average 3630.000
16:17:40 INFO | latency_cpu_min_average 3516.000
16:17:40 INFO | latency_cpu_max_average 165524.000

..and this can be automatically scraped and the results can be
instered into an influxdb for plotting later.

Signed-off-by: Colin Ian King <colin.king@canonical.com>