UBUNTU SAUCE: ubuntu_ltp_syscalls: use haveged for getrandom02
BugLink: https://bugs.launchpad.net/bugs/1831229

We tried to increase the timeout LTP_TIMEOUT_MUL for getrandom02:
  * c23ddd13271c12143d9f2da8eb71637d836cdc3d
  * 2b18c1a3f5311a979d83f84e601fae31bc47f123

But sometimes the entropy is still not enough.
Use haveged to generate more entropy for it, we need to exclude Trusty
as this package is only available since xenial.

The LTP_TIMEOUT_MUL algorithm should be left intact for Trusty.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>