UBUNTU SAUCE: ubuntu_ltp_syscalls: skip fanotify07 test for older kernels
BugLink: https://bugs.launchpad.net/bugs/1775165

On Ubuntu Xenial (kernel 4.4.0), test fanotify07 will fail with timeout
and cause hang, all the following fanotify* tests will fail with
timeout error because of this.

This is a valid regression-tester, but it's generating a false positive
result in the test report for other fanotify tests.

Matthew Ruffell tried to backport the fix for fanotify07 into upstream
4.4 and 4.9, but the scenario is too complicated to get this fixed, in
the end Greg K-H agreed with Matthew, the patches were dropped, and
this won't be getting fixed upstream.

So let's blacklist this test for older kernels.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>