UBUNTU_SECCOMP Update source to use upstream libs
The Current ubuntu_seccomp test is using a git repo that's 4 years old without any maintenance. The repo specifically tests seccomp_filter it would appear this test also fails to build on most clouds.

The following proposal is to update the test itself, using source from libseccomp, which not only includes tests for seccomp_filter but a wide array of others as well.

Signed-off-by: sfeole <sean.feole@canonical.com>