UBUNTU SAUCE: split xfstests into 3 different tests
BugLink: https://bugs.launchpad.net/bugs/1792348

When a test for a specific filesystem crashed while running the
xfstests, the test partition will be occupied until the system gets
rebooted. This will block the next test for a different filesystem, and
it is not only time wasting but causing unnecessary test failures.

Split the one huge chunk test into different tests against a specific
filesystem can solve this issue.

It will be splited into:
  * ubuntu_xfstests_btrfs
  * ubuntu_xfstests_ext4
  * ubuntu_xfstests_xfs

All of these are just a clone of xfstests, with filename / filesystem
name changed accordingly.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>