UBUNTU: SAUCE: ubuntu_zfs_xfs_geneirc: increase timeout, remove dedup and copies
We're seeing some test failures on some architectures because of
low memory making tests run a bit slower.  Bump the per-test timeout
to 40 minutes and disabled the slower dedup and copies settings. The
dedup and copies is tested with other zfs regression tests, so it
makes sense to disable them here as it only slows down the testing
and it's not really what we are focused on testing with these xfs
tests on ZFS.

Signed-off-by: Colin Ian King <colin.king@canonical.com>