UBUNTU SAUCE: ubuntu_zfs_xfs_generic: sync with upstream and fix gettid issue on Eoan
BugLink: https://bugs.launchpad.net/bugs/1844551

This test is using xfstest test suite, so just like ubuntu_xfstests_*,
sync with the latest upstream code to get rid of local hacks.

And with the latest upstream xfstest-bld, the gettid conflict issue for
newer glibc on Eoan can be fixed too.

The local patch to add zfs support must be updated as well.

Add libssl-dev as it's needed for the compilation.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>