UBUNTU: SAUCE: ubuntu_ltp_syscalls: fix fanotify09 for kernels older than 4.15
BugLink: https://bugs.launchpad.net/bugs/1804594

Test case #2 was added to fanotify09 as a regression test for the
following upstream commit:

b469e7e47c8a fanotify: fix handling of events on child sub-directory

which has been applied to v4.20 and backported to linux-4.19.y upstream
stable. We backported these changes from 4.19.y to 4.15 but they are not
trivial to backport to older kernels, so the best option is to patch the
test file to skip this testcase for older kernels.

While at it, fix a indentation issue found in testcase_blacklist().

Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>