UBUNTU SAUCE: ubuntu stress smoke test: disable clone stressor
This stressor can unpredictably trip the OOM killer on the autotests
which ends up in an aborted termination.  Since the OOM killer
is behaving unpredictably across different kernel versions the
best strategy for the moment is to drop the clone stressor. We've
tried different methods to stop the OOM killer doing this in the
past but the workarounds are not sufficient.

Signed-off-by: Colin Ian King <colin.king@canonical.com>