UBUNTU: SAUCE: ubunut_ltp: disable ntfs
BugLink: https://bugs.launchpad.net/bugs/1842266

The fs_fill test in fs will fail on NTFS with EROFS as well, disable
NTFS support like what we did for ubuntu_ltp_syscalls in e544f2a5.

This is making it harder for us to find out real regressions on
filesystems we really care about. As NTFS has never passed these tests,
disabling it won't be a problem in that respect.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>