Revert "UBUNTU: SAUCE: blacklist tests for trusty"
These two tests:
  - ubuntu_quota_smoke_test
  - ubuntu_zram_smoke_test

Are not working on Trusty AWS as we don't have linux-modules-extra-aws
package.

But it can still be tested on other flavour / kernels on Trusty.
Blacklisting them will drop regression coverage on these features.

A better solution is to build a linux-modules-extra-aws package or at
least just blacklist them on T-AWS.

This reverts commit a54839d364ff085a28e4826a669ed840234dff4b.