UBUNTU: SAUCE: ubuntu_performance_lkp, sockets
Update the way in which we implement the use of sockets
socket.gethostbyname(socket.gethostname()) will return,
127.0.0.1 if the host can't resolve DNS. This is a problem
with some offsite labs. Also, squid.internal is unknown
to our offsite labs.

The update will try to ssh to itself and return the default
routable interface.

The offsite lab in question is using 192.168 subnet so
no change is required there.

Signed-off-by: sfeole <sean.feole@canonical.com>