UBUNTU: SAUCE: ubuntu_performance_lkp: Conditionalize lkp-tests commit to use
The testing performed on akis needs to use a specific commit in the
lkp-tests tree but other machines should use a newer commit where things
are known to work better on the current Ubuntu development release
(Focal).

The upstream lkp-tests commit selected for the non-akis case includes a
handful of recent fixes for newer Ubuntu releases.

Suggested-by: Colin Ian King <colin.king@canonical.com>
Signed-off-by: Tyler Hicks <tyhicks@canonical.com>