UBUNTU: SAUCE: ubuntu_performance_lkp: Always enter the lkp-tests git tree
We should always change to the lkp-tests directory. One example of why
this is needed is the 'make install' command that is executed next.

Additionally, always reset the git tree to the specified commit by
passing '-B' rather than '-b' to git checkout. Using '-b; failed to
reset an existing "stable" branch if tests were ran back-to-back and
used different commits for the "stable" branch.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>
[PHLin: adjust context as we're not dropping bsdtar hack]
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>