UBUNTU: SAUCE: ubuntu_performance_lkp: Run hackbench atomic jobfiles
We should be running 'lkp split' on the keys of the dict and then 'lkp
run' on the resulting atomic jobfiles that we whitelist in the
corresponding values of the dict.

hackbench.yaml is not an existing jobfile so it should not be listed as
a key in the dictionary. Replace it with jobfiles that should be passed
to lkp split. The new values in the dictionary are the atomic jobfiles
that represent the tests that were being run when passing the non-atomic
jobfiles to lkp run.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>