UBUNTU: SAUCE: remove python-xattr dependency
Starting with Python 3.3, handling of Linux extended attributes is now
part of the Python Standard Library (os module).

On top of that, there isn't a single instantion of the xattr class, nor
any invokation of any xattr function in the entire autotest-client tree,
so i think this dependency is completely bogus and could be removed
altogether: i ran the patched ubuntu_xfstests_ext4 test in a Xenial VM
and in a Focal VM, and both completed fine.

Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>