UBUNTU: SAUCE: ubuntu_lxc: remove the reboot container after test
BugLink: https://bugs.launchpad.net/bugs/1788574

In some cases, the lxc-test-api-reboot test will left a reboot
container behind, and the next round of the LXC tests will fail because
of the existence of this container (normally happens on s390x zVM/LPAR,
as we don't re-deploy them).

Remove the reboot container after the test to solve this issue.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>