UBUNTU: SAUCE: add zfs encryption performance tests
Add 6 different ZFS encryption performance tests, each measuring 10
iterations of write and reads of data in a ram disk.  Putting the data
in a ramdisk allows us to factor out the noise of SSD or HDD backing
storage to get a better idea of the encryption overhead.

Output is in a machine parsable format that can be machine scraped and
dumped into influxdb for graphing, for example:

16:39:09 DEBUG| [stdout] zfs_encryption_aes_192_gcm_wr_duration_seconds_average 6.6886
16:39:09 DEBUG| [stdout] zfs_encryption_aes_192_gcm_wr_rate_mbytes_per_second_average 78.3800
16:39:09 DEBUG| [stdout] zfs_encryption_aes_192_gcm_wr_user_time_average .3110
16:39:09 DEBUG| [stdout] zfs_encryption_aes_192_gcm_wr_kern_time_average 4.7200
16:39:09 DEBUG| [stdout]
16:39:09 DEBUG| [stdout] zfs_encryption_aes_192_gcm_rd_duration_seconds_average 7.1901
16:39:09 DEBUG| [stdout] zfs_encryption_aes_192_gcm_rd_rate_mbytes_per_second_average 72.9200
16:39:09 DEBUG| [stdout] zfs_encryption_aes_192_gcm_rd_user_time_average .3060
16:39:09 DEBUG| [stdout] zfs_encryption_aes_192_gcm_rd_kern_time_average 6.8660

Signed-off-by: Colin Ian King <colin.king@canonical.com>