UBUNTU SAUCE: ubuntu_performance_fio will not build in 18.04
Commit #81c37704 Added a patch to fix some build errors in 18.10+
The patch renamed gettid which resolved a conflict between gettid
in glibc and fio.  Unfortunately this patch broke building FIO on
older series like 18.04. The following change simply adds a rule
to only apply the patch if the host machine != bionic.

Signed-off-by: Sean Feole <sean.feole@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>