UBUNTU: SAUCE: ubuntu_bpf: use clang-10 and llvm-10 for F-5.6
BugLink: https://bugs.launchpad.net/bugs/1879360

The ubuntu_bpf test can't be built with clang-9 on 5.6 kernel.
Use clang-10 and llvm-10 instead to solve this problem.

Also we still need to keep clang-9 for 5.4 kernel in Focal,
otherwise the test_map test will fail to execute on s390x with:
  libbpf: Error loading BTF: Invalid argument(22)

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>