UBUNTU: SAUCE: ubuntu_performance_iperf3: Fix collation of multi-instance stats
iperf3 doesn't support threads, so instead we run multiple clients and
collate the individual statistics. The collation part is broken. While our
test successfully manages multiple iperf3 client/server pairs, we tell each
one to log its stats to the same file. The code suggests an expectation
that each process would append its output to that file. But, iperf3 truncates
the file before writing, so we end up with only the output of the last process.
(And let us ignore the fact that even if that weren't so, multiple writers
to the same file at the same time is the bumpy road to garbletown). End
result being that a test that was supposed to sum up the totals of 8 different
clients was only really reporting the total of 1 of them.

Fix this by removing --logfile as a shared parameter and instead dynamically
create a unique temp file for each process and process them one-by-one.

Signed-off-by: dann frazier <dann.frazier@canonical.com>
Acked-by: Juerg Haefliger <juergh@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>