UBUNTU: SAUCE: ubuntu_performance_iperf3: remove incorrect comment
A comment asserts a spec 1.2 requirement that we achieve 90% of
(line) rate. No such requirement exists. The referenced spec only
provides criteria for flagging a change in performance vs. a previously
achieved baseline - it says nothing about how close that needs to be
to line rate.

I'm leaving the 90% check for PASS/FAIL in place, as it maybe useful in
other contexts. But it should be ignored for DGX2 regression testing
purposes.

Signed-off-by: dann frazier <dann.frazier@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>