UBUNTU: SAUCE: ubuntu_performance_iperf3: Emit per-iperf3-instance stats
To simplify debugging, and in the interest of transparency, I'd like to
provide as much raw data in our reports as possible, and expose the math we
use to formulate summaries (using spreadsheet formulas, in my case).
For multi-client iperf3 tests, we currently provide a sum of the individual
client rates, but not the individual rates themselves. We therefore aren't
able to say if, for example, an 8 client run regressed because one of
the clients gave us bogus data.

Add new fields to also log the individual rates.

Signed-off-by: dann frazier <dann.frazier@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>