UBUNTU: SAUCE: ubuntu_performance_iperf3: Use sudo when executing cpupower
While we use sudo in other places that need root privileges, we fail
to do so when setting the CPU frequency governor to performance. This
gives us different behavior when executed as root and non-root. This went
easily unnoticed because we redirect stderr to /dev/null, otherwise we
would've seen the error "Subcommand frequency-set needs root privileges"

Use sudo, and only redirect stdout to /dev/null.

Signed-off-by: dann frazier <dann.frazier@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>