UBUNTU: SAUCE: ubuntu_performance_iperf3: Bail if protocol != TCP
While we'll try and run the tests no matter what, we only have code
for reporting metrics in TCP mode. Instead of reporting nothing,
let's report an error if we find we're not in TCP mode. This saves
su a needless level of indention where all the reporting code
was previously in a 'if [ "${protocol}" == "TCP" ]' block.

It would be nice to not run the tests at all in !TCP mode, but the way
this test passes through iperf3 options makes that difficult to detect
until we have the command output to process.

Signed-off-by: dann frazier <dann.frazier@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>