UBUNTU: SAUCE: ubuntu_performance_iperf3: Run 3 iterations
According to the DGX test plan spec, each test is required to run 3 times,
but currently we only run the iperf3 tests once. Variation between these
runs is supposed to be reported, but with 1 run, that is now always 0.

Note: The min/max/avg calculations in this script don't appear to mean
what they do in other tests. Currently we appear to be comparing the
rates of individual clients in a single iteration, vs. the aggregate rate
of all clients across iterations. We currently only care about the
aggregate rates. I didn't adjust these calculations because I plan
to use a spreadsheet to do so - but perhaps we should consider dropping
those metrics lest their meaning cause confusion.

Signed-off-by: dann frazier <dann.frazier@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>