UBUNTU: SAUCE: ubuntu_kvm_smoke_test: use trap to do cleanup
If the KVM instance was created but the test did not complete
successfully, the script will terminates with exit 1 before the KVM can
be destroyed.

It is okay on maas-provisioned systems, but on a manual provisioned
instance like our s390x, you will see the following error message when
you try to run this test again:
  uvt-kvm: error: libvirt: storage volume 'bjf-test.qcow' exists already

Add a cleanup function with trap to make sure the KVM will be destroyed
when the script terminates.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>