UBUNTU: SAUCE: Improve the flavour parser
BugLink: https://bugs.launchpad.net/bugs/1888496

Some kernel will have flavour string with a dash in it,
for example 4.15.0-2006-azure-fips

In this case the string we get with split will be just "fips" and
making extra module installation not working.

Improve the flavour string parsing with regex. And add azure-fips to
the list.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>