UBUNTU: SAUCE: ubuntu_unionmount_ovlfs: add cleanup() for the test
Add cleanup() to unmount and remove the directory that we created for
the test.

The mkdir for /lower and /upper must be moved from setup() to
initialize() as setup() will be executed only once before the test
source code got copied to autotest/client/tmp/. After that if you try
to re-run the test, it won't be executed again.

Move them to initialize() to make sure the test directory will be
created on every run.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>