UBUNTU: SAUCE: ubuntu_kernel_selftests: disable memory hotplug for ARM
Memory hotplug test will fail on arm platforms for Linux 5.6+. It
was enabled in 5.6 but needs memory that does not have boot time pages
in the regions to be offlined and current test hardware cannot guarantee
that constraint so disable it for arm platforms for now.

Quoting Anshuman Khandual from ARM:

"A memory block should have these following conditions for it to be hot
removed on arm64 platform:

1. It needs to be removable

	- ZONE_MOVABLE areas are always removable
	- Other zone areas can be removable

2. It could be offlined

	- Must migrate out when attempted, page usage state should not block
	- ZONE_MOVABLE should work always, as its used from the user space
	- Kernel areas cannot be migrated, hence cannot be offlined and removed

3. It should not be boot memory

	- This is arm64 specific restriction
	- Must have been added after the boot, before being offlined or removed

The hot remove failed while trying to first offline the range because there
were unmovable pages. These are marked 'reserved', hence would assume that
they were used by the kernel. But even if kernel was not using those memory,
offline request would have still failed due to the third restriction above
i.e boot memory. The memory can be offlined only if it was added after boot."

Signed-off-by: Colin Ian King <colin.king@canonical.com>