UBUNTU: SAUCE: ubuntu_ltp_syscalls: compatibility fix for the absence of python2-yaml module on newer releases
On newer releases like groovy, the python2-yaml module does not exist
anymore.

Use the good old python dictionary to solve this issue.

The downside of this is that we will need to maintain two databases,
but I think it is ok for now since it's not changing constantly.

And eventually we could migrate to either one of them in the future.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>