UBUNTU: SAUCE: Ignore command status for DKMS installation
In ubuntu_lttng_smoke_test and ubuntu_sysdig_smoke_test we will check
the status of the DKMS module with "dkms status" command, however
when the installation command fails, the test will be interrupted and
run cleanup() on exit.

The "dkms status" will never get executed, ignore the installation
command status to solve this issue, let the DKMS make log can be
printed when it fails.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>