UBUNTU: SAUCE: ubuntu_fan_smoke_test: properly generate override.conf
BugLink: https://bugs.launchpad.net/bugs/1896511

To workaround a docker issue the following commit has been applied:

 433215fa "(UBUNTU: SAUCE: ubuntu_fan_smoke_test: workaround docker 'dependency' failure)"

However, with this fix, we are not properly generating override.conf,
that contains an extra "-e" at the beginning of the file, for example:

 ubuntu@moe:~$ cat /etc/systemd/system/containerd.service.d/override.conf
 -e [Service]
 ExecStartPre=

The reason is that the test script is using /bin/sh in the shabang (that
is /bin/dash) and dash'es echo doesn't support "-e".

Prevent this problem (and also other potential dash issues in the
future) by switching to /bin/bash.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>