UBUNTU: SAUCE: ubuntu_kernel_selftests -- pass KDIR to bpf selftest build
Linux 5.11 added a kernel module which is built as part of the
bpf selftest build. This fails to build when running our testing
because the source tree has not been configured and prepared for
module building. The linux-headers packages give us a tree
already set up for module builds, and the makefile will use
a different base for the build if supplied via the KDIR variable.
Point this variable ad the linux-headers install location for the
running kernel. This should not impact older releases as before
5.11 KDIR is not used in the bpf selftest makefiles.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>