UBUNTU: SAUCE: ubuntu_performance_pts.py: Bump John the Ripper version to 1.7.2 for Blowfish
The Blowfish test in John the Ripper 1.6.2 did not support OpenSSL 1.1,
making it incompatible with Ubuntu 20.04 and later. Bump the version
to 1.7.2 to support it. Note that while 1.7.2 does support earlier
versions of Ubuntu (I tested on 18.04), I've found that 1.6.2 and 1.7.2
produce uncomparable numbers, so consider this an epoch.

This does not change the version of John the Ripper used to run the
Traditional DES test. This is because John the Ripper dropped the
Traditional DES test before releasing 1.7.2. I found no interim release
of the John the Ripper PTS test that supports both DES and OpenSSL 1.1.

Signed-off-by: dann frazier <dann.frazier@canonical.com>
Acked-by: sean feole <sean.feole@canonical.com>