UBUNTU: SAUCE: ubuntu_ltp: split those passed tests to ubuntu_ltp_stable
Even with syscalls test split into the ubuntu_ltp_syscalls suite,
the ubuntu_ltp suite is still a huge chunk.

Those test that always pass will now be put into the
ubuntu_ltp_stable suite.

On the other hand, those subsets that keeps failing will be tested in
ubuntu_ltp, subsets as follows:
 * fs
 * mm
 * controllers
 * commands
 * cpuhotplug
 * net.ipv6_lib
 * cve
 * crypto
 * kernel_misc

In this way the ubuntu_ltp will be eaiser to review, and reduce the
chance to timeout on some instances.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>