ubuntu_cve2_kernel: CVE-2021-27363, CVE-2021-27364, CVE-2021-27365
This introduces tests for the above CVEs. On success, it will return 0 and only
output to stdout. On failure, it will return 1 and some output will go to
stderr. In case of kernels that are not mitigated, it tries not to crash them,
but that might still happen in some cases.

This will test whether an unprivileged user can read the iscsi transport handle
from sysfs, whether it can create an iscsi session, and whether a privileged
user can set a initiator name larger than page size.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>