UBUNTU SAUCE: ubuntu_zfs_xfs_generic: deprecate tests 123, 314, 317
These three tests require su'ing to the fsgqa xfs test user, however,
this is not working on some systems because the "Password:" prompt
is randomly appearing in the output stream causing the test case
pattern matching to fail randonly. Regex pattern matching does not
solve this issue because of the interleaved stdout streams, so
disable these tests as they are causing more hassle than they are
worth test-wise.

Signed-off-by: Colin Ian King <colin.king@canonical.com>