UBUNTU: SAUCE: ubuntu_boot: use kernel_taint_test script from checkbox
Use the well-written kernel_taint_test.py script taken from
plainbox-provider-checkbox [1]. It can translate the tainted flag
number into a human readable output, i.e. for the failure found in
s390x LPAR (lp:1892124) the report was like:
  ERROR: kernel tainted flag != 0: 4097

With this script:
  Kernel taint value is 4097
  Taint bit value: 0 (proprietary module was loaded)
  *   Modules with GPL Incompatible Licenses:
       zfs: CDDL
       zunicode: CDDL
       zavl: CDDL
       icp: CDDL
       zcommon: CDDL
       znvpair: CDDL
  Taint bit value: 12 (externally-built ('out-of-tree') module was loaded)
  *   Modules not in-tree:
       zfs
       zunicode
       zavl
       icp
       zcommon
       znvpair
       spl

It will be easier for tester to know what's going on here with just
a glance.

[1] https://git.launchpad.net/plainbox-provider-checkbox/

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>