UBUNTU: SAUCE: ubuntu_kernel_selftests: remove workaround for ftrace/kprobe_args_user.tc
This kprobe_args_user.tc in ftrace was disabled due to lp:1869227 it
will be marked as unresolved if not supported and marked as a failure.

With patch b730d668138cb3 ("ftrace/selftest: make unresolved cases
cause failure if --fail-unresolved set") landed (lp:1877958), an
unresolved case will not be treated as a failure unless we requested.
Therefore we can remove the workaround here.

Also note that with patch 1e11b7dbef17bb ("selftests/ftrace: Return
unsupported for the unconfigured features") this test will now be
marked correctly as unsupported.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>