UBUNTU: SAUCE: ubuntu_sysdig_smoke_test: fix test missing file names
The sysdig is unable to report the name of file used in reads and
writes.  This maybe worked some time ago, but definitely does not work
with sysdig 0.27.1.

For dd, cat or cp, the sysdig reports only arguments:
  cp (3558) < execve res=0 exe=cp args=/tmp/somefile./tmp/otherfile. tid=3558(cp) pid=3558(cp) ptid=1151(bash)

and all further reads/writes are to file descriptors, e.g.:
  write fd=4 size=131072

Also sysdig reporting tools fail to find the name of file used in IO:

  $ sysdig -r ${TMPFILE}.raw -c fdbytes_by fd.name
  Bytes               fd.name
  ------------------------------------------------
  1024.00KB           /dev/pts/1

  $ sysdig -r ${TMPFILE}.raw -c echo_fds
  ------ Write 1024B to   /dev/pts/1 (dd)

Fix the test by looking only for reads or writes tgo any descriptor.  It
is clearly a limitation of sysdig so the test should not expect and test
more than the tool can provide.

This fixes the test failures like:

  FAILED (trace at least 25 reads of /dev/zero by dd)
  FAILED (trace at least 25 writes to /dev/null by dd)

BugLink: https://bugs.launchpad.net/bugs/1844493
Signed-off-by: Krzysztof Kozlowski <krzysztof.kozlowski@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Guilherme G. Piccoli <gpiccoli@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>