UBUNTU: SAUCE: ubuntu_sysdig_smoke_test/ubuntu_lttng_smoke_test: always try to install DKMS when test starts
BugLink: https://bugs.launchpad.net/bugs/1931338

The DKMS package for these 2 tests will be removed with cleanup() in
the end of the test. This is for our manually provisioned system (bug
1798295), and it's also a good practice to keep the system clean after
the test.

However, since we only run setup() once if the test version is not
bumped, when you try to restart these tests on the same system without
removing autotest/client/tmp/<test suite> directory, they will fail
due to the lack of corresponding modules.

Solution for this is to move code in setup() into initialize(), which
is guaranteed to be executed on every run.

This is the workflow for a test:
1. initialize()
2. setup()
3. run_once()
4. cleanup()

With a second run immediately after the first one:
5. initialize()
6. run_once()
7. cleanup()

Code tested on an Azure instance and a bare-metal node.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Krzysztof Kozlowski <krzysztof.kozlowski@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>