UBUNTU: SAUCE: ubuntu_lttng_smoke_test: adjust memory limit to fix OOM
The test was failing on Azure Standard_B1ms instance due to OOM killer:

    utils:0153| [stdout] gcc: fatal error: Killed signal terminated program cc1

In journalctl:

    kernel: cc1 invoked oom-killer: gfp_mask=0x100cca(GFP_HIGHUSER_MOVABLE), order=0, oom_score_adj=0
    ...
    kernel: Free swap  = 0kB
    kernel: Total swap = 0kB
    kernel: 524174 pages RAM

The test sets up a swap on low-memory machines, however the detected
total memory on Standard_B1ms was 1.878 GB, so slightly above the 1.8
threshold and the swap was not setup.

Bump the threshold to really higher value, so this lack of memory won't
hit us.  Setting up swap (when not needed) should not impact the test,
except making setup() slightly slower.

BugLink: https://bugs.launchpad.net/bugs/1847989
Signed-off-by: Krzysztof Kozlowski <krzysztof.kozlowski@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>