UBUNTU: SAUCE: ubuntu_performance_deep_learning: init deep learning framework performance test
The purpose of this test is to generate the performance data of deep
learning framework. Currently it supports TensorFlow testing only.

If the target shell script complete to run and the target data file is
generated, the job is passed.

This test job is able to run and pass on a machine with expected devices
like a GPU supported by the expected drivers and deep learning
frameworks.

The test environment is mostly prepared by MAAS via a customized curtin
preseed. For tasks like driver installation, software installation
highly associated with driver or required reboot, are setup by the
preseed. The rest of tasks are completed by the autotest framework, and
defined in the corresponding testing job.

Signed-off-by: Taihsiang Ho (tai271828) <taihsiang.ho@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>