UBUNTU: SAUCE: ubuntu_ltp*: use our mirror for LTP
BugLink: https://bugs.launchpad.net/bugs/1935790

It takes time to get our patch applied on upstream, so let's use our
own mirror and put our patch on top of it as SAUCE patches.

By doing so we don't need to apply our local patches on the fly.
And we can have the same SAUCE patchset in all these 3 tests. The only
overhead is to keep the branch up-to-date.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>