UBUNTU: SAUCE: ubuntu_ltp_syscalls: use special branch for ESM series
BugLink: https://bugs.launchpad.net/bugs/1935790

Now we have our own mirror for LTP, we can use a dedicated branch for
Trusty and Xenial ESM series.

The local patch to skip the second test case in fanotify09 should be
applied on to the sru-trusty branch. Note that it might need to be
reworked as it will terminates the test early, there are five tests
in it, only one will be tested.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>