UBUNTU: SAUCE: ubuntu_qrt_kernel_hardening: resolve lengthy git clones
BugLink: n/a

The ubuntu_qrt_kernel_hardening test has been failing during SRU testing
due to the the qa-regression-testing repo taking over 20+ minutes to
clone. In many cases, this will result in a timeout and cause the test
to fail.

The Following changes are to address a new feature added to ckct.
SUT-TEST will rsync+ssh the qa-regression-testing to the SUT, prior to
executing autotest.

If the qa-regression-testing directory exists in ($HOME/), then use
that as the source. If not then we git clone qa-regression-testing
as we have always done so.

Signed-off-by: Sean Feole <sean.feole@canonical.com>