UBUNTU: SAUCE: ubuntu_qrt*: use pre-rsync git repo for the test
Experiment shows that it will take about 1 ~ 20 minutes to clone the
qa-regression-testing repo for our *qrt* tests, it's all depends on
the connection speed to the launchpad. And it seems quite unstable.

Take ubuntu_qrt_kernel_aslr_collisions on Oracle VM.Standard2.1 for
example, the jenkins job took 35 minutes to complete, but this
lengthy git clone task occupied 1/3 of the time.

It just need ~30 seconds to copy it from our jenkins server to a SUT
with rsync. I think it's worthy to optimize this for the rest of qrt
related tests like what we did for ubuntu_qrt_kernel_hardening test
in commit 6441fc1e9b ("UBUNTU: SAUCE: ubuntu_qrt_kernel_hardening:
resolve lengthy git clones")

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>