UBUNTU: SAUCE: ubuntu_bpf: stop the test when the build fails
BugLink: https://bugs.launchpad.net/bugs/1938663

If the setup job in ubuntu_bpf has failed, it should be stopped at
this point. Otherwise it will try to run the test_verifier and
test_maps sub-tests. These two will try to build the test again and
fail with the same reason.

By doing so we can simplify the hinting task, there will be just one
setup test case in the result page, which will be the only thing we
need to hint.

Tested on a KVM ndoe with OEM kernel and this works as expected.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>