UBUNTU: SAUCE: ubuntu_boot: implement revocation list checks
Implement revocation list checks. If kernel supports revocation lists,
check that 2012 canonical signing key is revoked.

Most kernels will skip this test reporting NA result, those kernels
that have support for revocation lists will check that it is correctly
configured and visible at runtime.

It is intentional for this to be part of ubuntu_boot test - kernels
failing this check must not be signed.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>