UBUNTU: SAUCE: ubuntu_lxc: override GPG server
BugLink: https://bugs.launchpad.net/bugs/1916087

These tests are known to fail with "ERROR: Unable to fetch GPG key
from keyserver":
  * lxc-test-apparmor-mount
  * lxc-test-autostart
  * lxc-test-unpriv
  * lxc-test-usernic
  * lxc-test-utils

It's because of the old GPG server is not responding anymore, and
this issue has been addressed in the lxc package[1].

This issue does not happen on Bionic and later release as we're
already using the updated exercise script from lxc package. Copy the
code snippet from the exercise script in lxc debian source package to
our script to solve this issue.

Note that with this patch applied, some of them will fail with:
  ERROR: Couldn't find a matching image.

This will be tracked in bug 1939537

Ref: https://bugs.launchpad.net/ubuntu/+source/lxc/+bug/1937945
[1] https://github.com/lxc/lxc/commit/f2a5d95d00a55bed27ef9920d67617cc75fecad8

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>